Fixing to run on python 3

+++ app/util/holiday.py
@@ -13,8 +13,8 @@ def easter_sunday(year):
 	k = c % 4
 	m = (a + 11 * h) / 319
 	r = (2 * e + 2 * j - k - h + m + 32) % 7
-	n = (h - m + r + 90) / 25 # mes
-	p = (h - m + r + n + 19) % 32 # dia
+	n = int((h - m + r + 90) / 25) # mes
+	p = int((h - m + r + n + 19) % 32) # dia
 
 	dt = datetime(year, n, p)
 	return dt
